Update: Added an inventory, and interactable items
Fixed previous inventory and loot system to work with the current project:
The inventory stores your current weapon and current tool, as well of a list of your consumables.
Interactables have a prompt for you to pick them up. Upon pressing E, the items represented by that interactable will be added to your inventory.
Each interactable can give you multiple consumables, and/or 1 tool and/or 1 weapon upon pick up. In the future it can be expanded to multiple weapons or tools.